Move *Buffer to serialization/buffer
