+++ bokehjs/src/lib/core/serialization/index.ts
@@ -1,2 +1,3 @@
 export {Serializer, SerializationError, Serializable, serialize} from "./serializer"
+export {Buffer, Base64Buffer} from "./buffer"
 export * from "./reps"
